feat(InventoryCardFormController): обновил метод "deleteInventoryCord", добавил проверку

+++ src/main/java/com/example/controllers/InventoryCardFormController.java
@@ -19,6 +19,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 
 import java.util.List;
+import java.util.Optional;
 
 @Controller
 @FxmlView("/com/example/fxml/inventory-card-form.fxml")
@@ -187,9 +188,20 @@ public class InventoryCardFormController {
     public void deleteInventoryCord() {
         InventoryCard selectedInventoryCard = tableView.getSelectionModel().getSelectedItem();
         if (selectedInventoryCard != null) {
-            inventoryCardService.deleteInventoryCard(selectedInventoryCard);
-            // Обновление отображения таблицы
-            updateTableView();
+            // Создание нового диалогового окна с подтверждением
+            Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+            alert.setTitle("Подтверждение удаления");
+            alert.setHeaderText(null);
+            alert.setContentText("Вы точно хотите удалить этот объект?");
+
+            // Отображение диалогового окна и ожидание решения пользователя
+            Optional<ButtonType> result = alert.showAndWait();
+            if (result.isPresent() && result.get() == ButtonType.OK) {
+                // Если пользователь подтвердил удаление, удалить объект
+                inventoryCardService.deleteInventoryCard(selectedInventoryCard);
+                // Обновление отображения таблицы
+                updateTableView();
+            }
         }
     }
